// src/app/models/sell.ts
export interface sell {
    transactionId: number;
    transDate: Date;
    nama: string;
    noLicense: string;
    noResit: string;
    letterOfConsentNo: String;
    type: string;
    sekerap: number;
    lateks: number;
    weight: number;
    drc: number;
    total: number;
    unitPrice: number;
}